requirements.txt
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-oauthlib==1.2.3
requests==2.32.5
yt-dlp==2026.03.03